% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot-sel.R
\name{plot_sel}
\alias{plot_sel}
\title{Plot Selectivity}
\usage{
plot_sel(
  Year = M$Year,
  sel = M[, 2:10],
  styr = 1964,
  fage = NULL,
  lage = NULL,
  alpha = 0.2,
  scale = 3.8,
  fill = "purple"
)
}
\arguments{
\item{Year}{A vector of years. Default is \code{M$Year}.}

\item{sel}{A matrix of selectivity values. Default is \code{M[,2:10]}.}

\item{styr}{An integer specifying the start year for the plot. Default is 1964.}

\item{fage}{An integer specifying the first age to include in the plot. Default is NULL.}

\item{lage}{An integer specifying the last age to include in the plot. Default is NULL.}

\item{alpha}{A numeric value specifying the transparency level of the fill. Default is 0.2.}

\item{scale}{A numeric value specifying the scale for the density ridges. Default is 3.8.}

\item{fill}{A character string specifying the fill color for the density ridges. Default is "purple".}
}
\value{
A ggplot object showing selectivity across years and ages.
}
\description{
This function creates a plot of selectivity across years and ages.
}
\examples{
\dontrun{
plot_sel(Year = M$Year, sel = M[,2:10], styr = 1970, fage = 1, lage = 8)
}

}
