% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-agefit.R
\name{PlotAgeFit}
\alias{PlotAgeFit}
\title{Plot Age Fit}
\usage{
PlotAgeFit(x = bc, title = NULL, type = "fishery", fage = 2, lage = 7)
}
\arguments{
\item{x}{A list containing the observed and predicted age compositions. Default is \code{bc}.}

\item{title}{A character string for the plot title. Default is NULL.}

\item{type}{A character string specifying the type of composition to plot. Default is "fishery".}

\item{fage}{An integer specifying the first age to include in the plot. Default is 2.}

\item{lage}{An integer specifying the last age to include in the plot. Default is 7.}
}
\value{
A ggplot object comparing observed and predicted age compositions.
}
\description{
This function creates a plot comparing observed and predicted age compositions for a specified type.
}
\examples{
\dontrun{
PlotAgeFit(x = bc, title = "Age Composition", type = "survey", fage = 1, lage = 8)
}

}
