% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_re_output}
\alias{read_re_output}
\title{Read output file from the random effects (RE) model into a
data.frame}
\usage{
read_re_output(file = "rwout.rep", skip_data = FALSE, use_names = FALSE)
}
\arguments{
\item{file}{A character string of the file to be read in,
defaulting to 'rwout.rep' in the current working directory.}

\item{skip_data}{Whether to skip returning columns
representing the data inputs. If TRUE these are included and
will have NA values for any missing years}

\item{use_names}{Whether to use the names written by the RE
model or to use more meaningful ones (default)}
}
\value{
A dataframe with columns named based on the tags in
the output file
}
\description{
Read output file from the random effects (RE) model into a
data.frame
}
