% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nh.R
\name{run_nh}
\alias{run_nh}
\title{Run NH Model}
\usage{
run_nh(model = "m1", steepness = 0.7, runit = TRUE)
}
\arguments{
\item{model}{A character string specifying the model directory. Default is "m1".}

\item{steepness}{A numeric value for the steepness parameter. Default is 0.7.}

\item{runit}{A logical value indicating whether to run the model. Default is TRUE.}
}
\value{
The function returns the output read from the \code{nh_R.rep} file in the specified model directory.
}
\description{
This function runs the NH model with specified parameters and returns the output.
}
\examples{
\dontrun{
m2 <- run_nh("m2")
}

}
