% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_re_species}
\alias{test_re_species}
\title{Test the RE on a single species. This compares a current ADMB
univariate run to a previous run. It tests \emph{continuity} not
correctness}
\usage{
test_re_species(d, test_tmb = TRUE)
}
\arguments{
\item{d}{The .dat file name, e.g., bigskate_t.dat, found in
tests/data folder}
}
\value{
A data.frame with ADMB and TMB estimates for comparing
}
\description{
Test the RE on a single species. This compares a current ADMB
univariate run to a previous run. It tests \emph{continuity} not
correctness
}
\details{
It will throw an error if the results have changed,
otherwise results are consistent.
}
