% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_re_input}
\alias{write_re_input}
\title{Write an input file for the RE model}
\usage{
write_re_input(
  file,
  years,
  biomass,
  CV,
  tag = NULL,
  sp = "species",
  survey = "survey",
  yr_start = NULL,
  yr_end = NULL
)
}
\arguments{
\item{file}{A path to the output file, usually ending in .dat}

\item{years}{A vector of years for the observations}

\item{biomass}{A vector of biomasses for the observations}

\item{CV}{A vector of CV values for the observations}

\item{tag}{Optional character string to print to file for
bookkeeping (e.g., "EBS shelf flathead sole")}

\item{yr_start}{Optional start year for predictions}

\item{yr_end}{Optional end year for predictions}
}
\value{
Nothing, a data file is written as \param{file}
}
\description{
Write an input file for the RE model
}
\seealso{
read_re_output
}
